rename: useSearchPlaces.ts를 searchShopsUsecase.ts로 이동
